% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_edits.R
\name{print_edit_summary}
\alias{print_edit_summary}
\title{Print Summary of Edit Comparison}
\usage{
print_edit_summary(comparison_df)
}
\arguments{
\item{comparison_df}{Output from compare_edits()}
}
\description{
Print Summary of Edit Comparison
}
